Fix FD_SETSIZE comparision warnings on FreeBSD
* And again, FreeBSD sucks... it defines FD_SETSIZE as an unsigned integer,
  while every fd is signed.
  This obviously breaks thinks like assert(fd < FD_SETSIZE);


git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@2564 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/fdevent_select.c
@@ -54,7 +54,7 @@ static int fdevent_select_event_del(fdevents *ev, iosocket *sock) {
 static int fdevent_select_event_add(fdevents *ev, iosocket *sock, int events) {
 	/* we should be protected by max-fds, but you never know */
 #ifndef _WIN32
-	assert(sock->fd < FD_SETSIZE);
+	assert(sock->fd < ((int)FD_SETSIZE));
 #endif
 
 	if (events & FDEVENT_IN) {
